Friday afternoon comments
Other comments:
'Minimally parsed' is problematic IMO. Make it clear what is and
is not part of the spec. 1.5.3 is a bit confusing in this regard.

Inconsistent formatting of phrase 'comment lines' throughout
e.g. 1.5.3

2.3.3 wording needs work, think tom has revised already

Feel free to normalise my erratic use of file format specification
vs interchange format, either or both is fine.